pyproject.toml
[tool.poetry]
authors = ["Cloves Almeida <almeida.cloves@bcg.com>"]
description = "Opinionated config handling for BCG Gamma projects"
name = "gamma-config"
packages = [
  {include = "gamma"},
]
version = "0.2.15"

[tool.poetry.dependencies]
colorama = "^0.4.4"
multimethod = {git = "https://github.com/cjalmeida/multimethod.git", branch = "main"}
python = "^3.6"
python-dateutil = "^2.8.1"
python-dotenv = "0.13.*"
"ruamel.yaml" = "0.16.*"

[tool.poetry.dev-dependencies]
black = "19.10b0"
bump2version = "1.0.*"
flake8 = "3.8.*"
httpx = "0.16.*"
jinja2 = "2.11.*"
pytest = "6.2.*"
sphinx = "3.2.*"
sphinx-autobuild = "2020.9.*"
sphinx-rtd-theme = "0.5.*"
pytest-cov = "^2.11.1"

[tool.pytest.ini_options]
addopts = "--cov-report xml:cov.xml --cov gamma"
testpaths = "test"
